Applying possible solution for undesired behavior.
When a stock symbol is added or removed, the change is applied immediately to the current view, however, when the app is restarted, none of the previous changes in stock list SharedPreference exist.

Reported as issue #32 on StockHawk repository.
"Symbol changes are not saving correctly to disk SharedPreferences"
Read for details: https://github.com/udacity/StockHawk/issues/32